+++ CHANGELOG.md
@@ -93,6 +93,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+
+
+## [4.185.0] - 2026-07-06
+
+### Added
+
+- feat(bio): ✨ more vision-verified photos (phase 1) (#563)\n- feat(bio): ✨ more vision-verified photos (phase 1) (#563)\n
+
+## [4.184.0] - 2026-07-06
+
+### Added
+
+- fix(bio): 🐛 restore generation + resolve hung UI (#562)\n- fix(bio): 🐛 restore generation + resolve hung UI (#562)\n
+
 ## [4.183.0] - 2026-07-05
 
 ### Added